refactor(client): AllocationManager -> NetworkAllocation
